+++ CHANGELOG.md
@@ -100,6 +100,7 @@
 - Dev: Fixed most compiler warnings. (#5028)
 - Dev: Added the ability to show `ChannelView`s without a `Split`. (#4747)
 - Dev: Refactor Args to be less of a singleton. (#5041)
+- Dev: Channels without any animated elements on screen will skip updates from the GIF timer. (#5042)
 
 ## 2.4.6